tumbls & tweets realiably
Updated the way the Tumbls and Tweets are gathered. All tumbls with the particular account will be gathered and diffed. 20 at a time, which is the API max. For now, there will be no delay between calls. But may need to down the road. Tweets are called for, and diffed. Since we can't realiably get all tweets, we will only store the ones that we can get. Which means we don't have any  permalinks to tweets on our news page.

+++ news/management/commands/worker_hash_tweets.py
@@ -7,7 +7,6 @@ from ...sources.hash_tweets import HashTweets
 
 from ...models import HashTweet as SteamHashTweet
 
-
 logger = logging.getLogger(__name__)
 
 
@@ -22,39 +21,43 @@ class Command(BaseCommand):
             # where we will store all tumbls
             tweets = HashTweets()
 
-            # get tumblr posts that are already
-            # represented in the database, and
-            # see if they need to be updated.
-            logging.info("Updating existing HashTweet posts")
-            existing_tweets = SteamHashTweet\
-                .objects\
-                .filter(tid__in=tweets.tids)
+            if (tweets.successful_connection):
 
-            for existing_tweet in existing_tweets:
-                # compare existing features to tweet data
-                # updates if necessary
-                tweets.compare(existing_tweet)
+                # get tumblr posts that are already
+                # represented in the database, and
+                # see if they need to be updated.
+                logging.info("Updating existing HashTweet posts")
+                existing_tweets = SteamHashTweet\
+                    .objects\
+                    .filter(tid__in=tweets.tids)
 
-            # get entries in the database
-            # that are not represented in the
-            # twitter data. means they have been
-            # deleted, and should be removed
-            # here as well
-            logging.info("Removing deleted HashTweet posts")
-            SteamHashTweet\
-                .objects\
-                .exclude(tid__in=tweets.tids)\
-                .delete()
+                for existing_tweet in existing_tweets:
+                    # compare existing features to tweet data
+                    # updates if necessary
+                    tweets.compare(existing_tweet)
 
-            # create new tumblr posts
-            logging.info("Figuring out which HashTweet to create")
-            tweets_to_create = tweets.to_create()
+                # get entries in the database
+                # that are not represented in the
+                # twitter data. means they have been
+                # deleted, and should be removed
+                # here as well
+                logging.info("Removing deleted HashTweet posts")
+                SteamHashTweet\
+                    .objects\
+                    .exclude(tid__in=tweets.tids)\
+                    .delete()
 
-            for tweet in tweets_to_create:
-                obj = SteamHashTweet(**tweet.data())
-                obj.save()
+                # create new tumblr posts
+                logging.info("Figuring out which HashTweet to create")
+                tweets_to_create = tweets.to_create()
 
-            logging.info("HashTweets now up to date.")
+                for tweet in tweets_to_create:
+                    obj = SteamHashTweet(**tweet.data())
+                    obj.save()
+
+                logging.info("HashTweets now up to date.")
+            else:
+                logging.info("Could not update Hash Tweets.")
 
         except CommandError as detail:
             err = 'Error getting HashTweet data! ' +\
